move postContainer creation code to new file
